refactor: extract spoilers and reuse
