+++ src/instruments/src/SD/Common/ComponentSidePositionProps.ts
@@ -1,0 +1,3 @@
+export interface ComponentSidePositionProps {
+    side: 'left' | 'right'
+}
