+++ src/instruments/src/SD/Common/Spoilers.scss
@@ -1,0 +1,34 @@
+@import "../../Common/definitions.scss";
+.Spoilers {
+  &.GreenShapeThick {
+    fill: transparent;
+    stroke: $display-green;
+    stroke-width: 3;
+  }
+
+  &.WarningShapeThick {
+    fill: transparent;
+    stroke: $display-amber;
+    stroke-width: 3;
+  }
+
+  &.GreenShape {
+    fill: transparent;
+    stroke: $display-green;
+    stroke-width: 2;
+  }
+
+  &.WarningShape {
+    fill: transparent;
+    stroke: $display-amber;
+    stroke-width: 2;
+  }
+
+  &.Warning {
+    fill: $display-amber;
+  }
+
+  &.Standard {
+    font-size: $font-size-larger;
+  }
+}
