+++ src/instruments/src/SD/Pages/Fctl/Fctl.scss
@@ -8,18 +8,6 @@
     fill: $display-white;
   }
 
-  .GreenShapeThick {
-    fill: transparent;
-    stroke: $display-green;
-    stroke-width: 3;
-  }
-
-  .WarningShapeThick {
-    fill: transparent;
-    stroke: $display-amber;
-    stroke-width: 3;
-  }
-
   .MainShape {
     fill: transparent;
     stroke: $display-light-grey;
